requirements.txt
Flask==0.10.1
Flask-MySQL==1.3
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.10.4
itsdangerous==0.24
pyparsing==2.0.3
wsgiref==0.1.2
stemming==1.0.1
dateutils==0.6.6